requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
reportlab==4.4.5
zxing-cpp==2.3.0
Pillow>=9.0.0
transformers>=4.30.0
torch>=2.0.0

